requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# MongoDB ODM
beanie>=1.23.0
motor>=3.3.0

# Settings and configuration
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4

# Google Gemini API (Free tier available)
google-generativeai>=0.3.0

# Additional utilities
email-validator>=2.1.0

# HTTP client for API calls
aiohttp>=3.9.0

# OpenAI API
openai>=1.3.0